refactor(qbittorrent-e2e): move verify_payload_integrity to scenario_steps
The function is an ASSERT step, not runner logic. It checks a meaningful
scenario postcondition (downloaded file matches original payload) and is
likely to be reused across future transfer scenarios.

Move it to a dedicated scenario_steps/verify_payload_integrity.rs file,
re-export it from scenario_steps/mod.rs, and import it in runner.rs.

Side effects in runner.rs:
- Remove the now-redundant verify_payload_integrity fn and sha1_hex helper
- Drop dead imports: std::fmt::Write, sha1::{Digest, Sha1}
